# api/openapi.yaml
openapi: 3.0.3
info:
  title: SignalWatch API
  version: 2.5.0-credentials
  description: |
    SignalWatch M4 HTTP API contract. All timestamps are UTC RFC3339 values.
    The public API covers identity, profile settings, the controlled source
    catalog, subscription CRUD, and the current user's matched arXiv papers.
    Background local matching provides explainable match reasons. M4 adds the
    subscription-scoped daily Digest (scheduled in the user timezone) and email delivery Worker; mail delivery operations
    are intentionally not exposed as public endpoints. User-supplied provider credentials provide optional paper summaries and cached Digest enrichment; Bounded Agents provide persistent subscription drafting and evidence-based paper conversations. Read-only operations endpoints expose safe runtime
    status to database-authorized operator accounts; no administrative mutation
    or log-download endpoint is exposed.
...
security:
  - bearerAuth: []
paths:
...
  /api/v2/subscriptions:
    post:
      tags: [Subscriptions]
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/SourceNotFound"
        "409":
          $ref: "#/components/responses/SubscriptionLimitReached"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Subscriptions]
      operationId: listSubscriptions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: source_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Stable, newest-first page of owned subscriptions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v2/papers:
    get:
      tags: [Papers]
      operationId: listMatchedPapers
      description: Returns each paper once, with all match reasons owned by the current user.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: subscription_id
          in: query
          description: Limit papers to one subscription owned by the current user.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Newest-first page ordered by first discovery time and paper ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: signalwatch_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token returned by the login endpoint.
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
        call_id: {type: string, maxLength: 64}
        retry_after_seconds: {type: integer, minimum: 0}
        reset_at: {$ref: "#/components/schemas/UTCTimestamp"}
      example:
        code: VALIDATION_ERROR
        message: request is invalid
        request_id: req_01j_example
...
    PublicSource:
      type: object
      additionalProperties: false
      required: [id, source_key, kind, name, rule_types, allowed_categories]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        source_key:
          type: string
          minLength: 1
          maxLength: 64
          example: arxiv
        kind:
          type: string
          enum: [arxiv]
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: arXiv
        rule_types:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/RuleType"
        allowed_categories:
          type: array
          uniqueItems: true
          items:
            type: string
          example: [cs.AI, cs.CL, cs.CV, cs.IR, cs.LG, cs.RO, cs.SE]
    RuleType:
      type: string
      enum: [category, include_keyword]
    Rules:
      type: object
      additionalProperties: false
      required: [category, keywords]
      properties:
        category: {type: string, minLength: 1, maxLength: 100}
        keywords:
          type: array
          maxItems: 30
          uniqueItems: true
          items: {$ref: "#/components/schemas/RuleValue"}
    CreateRules:
      type: object
      additionalProperties: false
      required: [category]
      properties:
        category: {type: string, minLength: 1, maxLength: 100}
        keywords:
          type: array
          maxItems: 30
          uniqueItems: true
          default: []
          items: {$ref: "#/components/schemas/RuleValue"}
    BackfillStatus:
      type: object
      additionalProperties: false
      required: [state, processed, matched]
      properties:
        state: {type: string, enum: [pending, processing, complete, failed, cancelled]}
        processed: {type: integer, minimum: 0}
        matched: {type: integer, minimum: 0}
    RuleValue:
      type: string
      minLength: 1
      maxLength: 100
      description: Trimmed; consecutive Unicode whitespace is folded for text rules.
    CreateSubscriptionRequest:
      type: object
      additionalProperties: false
      required: [source_id, name, rules]
      properties:
        max_items_per_digest:
          type: integer
          minimum: 1
          maximum: 20
          description: Independent per-subscription email limit. Creation defaults to the user's current default limit.
        source_id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
          minLength: 1
          maxLength: 100
        objective:
          $ref: "#/components/schemas/NullableObjective"
        enabled:
          type: boolean
          default: true
        digest_ai_enabled: {type: boolean, default: false}
        digest_ai_language: {type: string, enum: [zh, en], default: zh}
        rules:
          $ref: "#/components/schemas/CreateRules"
...
    Subscription:
      type: object
      additionalProperties: false
      required:
        - backfill
        - max_items_per_digest
        - digest_ai_enabled
        - digest_ai_language
        - id
        - source
        - name
        - objective
        - enabled
        - version
        - rules
        - created_at
        - updated_at
      properties:
        backfill: {$ref: "#/components/schemas/BackfillStatus"}
        max_items_per_digest:
          type: integer
          minimum: 1
          maximum: 20
          description: Independent per-subscription email limit. Creation defaults to the user's current default limit.
        digest_ai_enabled: {type: boolean}
        digest_ai_language: {type: string, enum: [zh, en]}
        id:
          type: integer
          format: int64
          minimum: 1
        source:
          $ref: "#/components/schemas/PublicSource"
        name:
          type: string
          minLength: 1
          maxLength: 100
        objective:
          $ref: "#/components/schemas/NullableObjective"
        enabled:
          type: boolean
        version:
          type: integer
          format: int32
          minimum: 1
        rules:
          $ref: "#/components/schemas/Rules"
        created_at:
          $ref: "#/components/schemas/UTCTimestamp"
        updated_at:
          $ref: "#/components/schemas/UTCTimestamp"
    SubscriptionPage:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
    PaperMatch:
      type: object
      additionalProperties: false
      required:
        - subscription_id
        - subscription_name
        - category
        - subscription_active
        - matched_keywords
        - matched_at
      properties:
        subscription_id:
          type: integer
          format: int64
          minimum: 1
        subscription_name:
          type: string
          minLength: 1
          maxLength: 100
        category:
          $ref: "#/components/schemas/RuleValue"
        subscription_active:
          type: boolean
        matched_keywords:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/RuleValue"
        matched_at:
          $ref: "#/components/schemas/UTCTimestamp"
    Paper:
      type: object
      additionalProperties: false
      required:
        - id
        - source_id
        - arxiv_id
        - title
        - abstract
        - comments
        - authors
        - categories
        - published_at
        - arxiv_updated_at
        - arxiv_url
        - pdf_url
        - first_seen_at
        - matches
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        source_id:
          type: integer
          format: int64
          minimum: 1
        arxiv_id:
          type: string
          minLength: 1
          maxLength: 64
        title:
          type: string
          minLength: 1
        abstract:
          type: string
        comments:
          type: string
        authors:
          type: array
          items:
            type: string
            minLength: 1
        categories:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            minLength: 1
        published_at:
          $ref: "#/components/schemas/UTCTimestamp"
        arxiv_updated_at:
          $ref: "#/components/schemas/UTCTimestamp"
        arxiv_url:
          type: string
          format: uri
        pdf_url:
          type: string
          format: uri
        first_seen_at:
          $ref: "#/components/schemas/UTCTimestamp"
        matches:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PaperMatch"
    PaperPage:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Paper"
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
    NullableObjective:
      type: string
      nullable: true
      maxLength: 500
      description: Blank input is normalized to null; nonblank input is trimmed.
    UTCTimestamp:
      type: string
      format: date-time
      description: UTC RFC3339 timestamp.
      example: "2026-09-05T08:30:00Z"
...
  responses:
...
    ValidationError:
      description: Request syntax, parameter, or business validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION_ERROR
            message: request is invalid
            request_id: req_01j_example
    Unauthorized:
      description: Missing, invalid, expired, or tampered bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: AUTH_UNAUTHORIZED
            message: unauthorized
            request_id: req_01j_example
...
    SourceNotFound:
      description: Source does not exist or is disabled.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: SOURCE_NOT_FOUND
            message: source not found
            request_id: req_01j_example
...
    SubscriptionLimitReached:
      description: The user already has 20 active enabled subscriptions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: SUBSCRIPTION_LIMIT_REACHED
            message: enabled subscription limit reached
            request_id: req_01j_example
...
    MethodNotAllowed:
      description: The path exists but does not support the request method.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: METHOD_NOT_ALLOWED
            message: method not allowed
            request_id: req_01j_example
    InternalError:
      description: Unexpected server failure; internal details are not exposed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INTERNAL_ERROR
            message: internal server error
            request_id: req_01j_example